+++ src/index.ts
@@ -3,3 +3,5 @@ export * from './Popover';
 export * from './Dialog';
 export * from './Drawer';
 export * from './Menu';
+export * from './ContextMenu';
+export * from './Notification';